# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Iniciar sesión de usuario",
                "parameters": [
                    {
                        "description": "Credenciales de inicio de sesión",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gemini/process": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Iniciar procesamiento de prompt",
                "parameters": [
                    {
                        "description": "Prompt y modelo a procesar",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PromptRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.GeminiProcessingIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gemini/process-file": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Iniciar procesamiento con archivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prompt",
                        "name": "prompt",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Modelo (opcional, por defecto gemini-3-flash-preview)",
                        "name": "model",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Archivo (pdf/png/jpg)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.GeminiProcessingFileIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gemini/status-file/{gemini_processing_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Obtener estado de procesamiento de archivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del proceso",
                        "name": "gemini_processing_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeminiProcessingFileResponse"
                        }
                    }
                }
            }
        },
        "/gemini/status/{gemini_processing_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Obtener estado de procesamiento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del proceso",
                        "name": "gemini_processing_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeminiProcessingResponse"
                        }
                    }
                }
            }
        },
        "/learning/chat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "learning"
                ],
                "summary": "Iniciar tutoría de conversación con IA",
                "parameters": [
                    {
                        "description": "Mensaje del estudiante y modelo opcional",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PromptRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.GeminiProcessingIDResponse"
                        }
                    }
                }
            }
        },
        "/learning/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "learning"
                ],
                "summary": "Obtener historial de aprendizaje",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LearningInteractionDB"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener todos los usuarios",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Crear usuario",
                "parameters": [
                    {
                        "description": "Datos para crear usuario",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/email/{email}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener usuario por Email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email del usuario",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Actualizar usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email del usuario",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos para actualizar usuario",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/email/{email}/language": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Actualiza únicamente el idioma objetivo y el nivel del usuario.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Actualizar configuración de idioma",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email del usuario",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de idioma",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateLanguageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener usuario por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Eliminar usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "description": "Opcional: Podrías incluir datos del usuario (nombre, email) aquí.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "efren@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Efren David"
                },
                "language_level": {
                    "type": "string",
                    "example": "A1"
                },
                "password": {
                    "type": "string",
                    "example": "miPasswordSeguro123"
                },
                "target_language": {
                    "type": "string",
                    "example": "English"
                }
            }
        },
        "models.GeminiProcessingFileIDResponse": {
            "type": "object",
            "properties": {
                "task_id": {
                    "type": "string",
                    "example": "8b9a1d2e-3c4f-5a6b-7c8d-9e0f1a2b3c4d"
                }
            }
        },
        "models.GeminiProcessingFileResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.GeminiProcessingStatus"
                }
            }
        },
        "models.GeminiProcessingIDResponse": {
            "type": "object",
            "properties": {
                "task_id": {
                    "type": "string",
                    "example": "8b9a1d2e-3c4f-5a6b-7c8d-9e0f1a2b3c4d"
                }
            }
        },
        "models.GeminiProcessingResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "8b9a1d2e-3c4f-5a6b-7c8d-9e0f1a2b3c4d"
                },
                "result": {
                    "type": "string",
                    "example": "Sí, existen varias becas..."
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GeminiProcessingStatus"
                        }
                    ],
                    "example": "finalizado"
                }
            }
        },
        "models.GeminiProcessingStatus": {
            "type": "string",
            "enum": [
                "pendiente",
                "en_proceso",
                "finalizado",
                "error"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusError"
            ]
        },
        "models.LearningInteractionDB": {
            "type": "object",
            "properties": {
                "conversationID": {
                    "description": "👈 NUEVO",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interaction_type": {
                    "description": "Ejemplo: Conversation, Grammar, Exercise",
                    "type": "string",
                    "example": "Correction"
                },
                "language": {
                    "type": "string",
                    "example": "French"
                },
                "level": {
                    "type": "string",
                    "example": "B2"
                },
                "prompt": {
                    "type": "string",
                    "example": "Write a dialogue about a train ticket."
                },
                "response": {
                    "type": "string",
                    "example": "Bonjour, je voudrais acheter un billet."
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Clave Foránea al usuario",
                    "type": "integer"
                }
            }
        },
        "models.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "efren@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "miPasswordSeguro123"
                }
            }
        },
        "models.PromptRequest": {
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "model": {
                    "type": "string",
                    "example": "gemini-3-flash-preview"
                },
                "prompt": {
                    "type": "string",
                    "example": "Conoces las becas para Finlandia?"
                }
            }
        },
        "models.UpdateLanguageInput": {
            "type": "object",
            "properties": {
                "language_level": {
                    "type": "string",
                    "example": "B2"
                },
                "target_language": {
                    "type": "string",
                    "example": "English"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "efren@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Efren David"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}